Stash fishing rod after tutorial
